backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning & Image Processing
tensorflow==2.15.0
keras==2.15.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
albumentations==1.3.1
imgaug==0.4.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# API & Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
aiofiles==23.2.1
